Create the framework code for managing the local database.
Created the new class DatabaseIO, which contains all the methods necessary for database access. Additionally, Javadocs for every the methods have been added. The code does not, however, contain the SQL prepared statements yet. This will be updated in a future commit, once the schema is determined.

MainActivity has been slightly modified to accommodate the database. Note that this code has not been tested live, as the SQL statements are not functioning.