Offline Login working, FCM configured
I have used shared preferences to store the credentials of the user to
initiate Offline login.

SHA 256 algorithm is used to hash the password before storing it. A
random salt is also added to the password before hashing it and stored
in the shared prefernce.

Offline login is possible only if there is no network

FCM is also configured , and it is possible to invalidate offline login
credentials by sending a push notification using FCM